+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/ManagementClientGenerator.cs
@@ -3,6 +3,7 @@
 
 using Azure.Generator.Management.Visitors;
 using Azure.Generator.Management.Providers;
+using Azure.Generator.Management.Utilities;
 using Azure.ResourceManager;
 using Microsoft.CodeAnalysis;
 using Microsoft.TypeSpec.Generator;
@@ -46,6 +47,9 @@ namespace Azure.Generator.Management
         /// <inheritdoc/>
         public override ManagementTypeFactory TypeFactory { get; }
 
+        private ResourceDataCustomizationResolver? _resourceDataCustomizationResolver;
+        internal ResourceDataCustomizationResolver ResourceDataCustomizationResolver => _resourceDataCustomizationResolver ??= new();
+
         /// <inheritdoc/>
         public override TypeProviderWriter GetWriter(TypeProvider provider)
         {
@@ -74,6 +78,7 @@ namespace Azure.Generator.Management
         /// </summary>
         protected override void Configure()
         {
+            AddCustomCodeAttributeProvider(OutputLibrary.CodeGenResourceDataAttributeDefinition);
             base.Configure();
             // Include Azure.ResourceManager
             AddMetadataReference(MetadataReference.CreateFromFile(typeof(ArmClient).Assembly.Location));